# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: Payment Gateway API - Enterprise Payment Processing
  description: |
    # 🚀 Payment Gateway API v1.0
    
    **Enterprise-grade payment processing API** with comprehensive Authorize.Net integration, designed for secure, scalable, and compliant financial transactions.
    
    ## 🎯 Core Capabilities
    
    ### 💳 Payment Processing
    - **Direct Purchase**: One-step authorization and capture
    - **Two-Step Processing**: Separate authorization and capture
    - **Transaction Management**: Void and refund operations
    - **Payment Method Validation**: Pre-transaction verification
    
    ### 📄 Subscription Management
    - **Recurring Billing**: Automated subscription processing
    - **Plan Management**: Flexible subscription plans with trials
    - **Lifecycle Control**: Pause, resume, and cancel subscriptions
    - **Dunning Management**: Intelligent failed payment handling
    
    ### 📊 Analytics & Reporting
    - **Real-time Dashboards**: Live transaction monitoring
    - **Business Intelligence**: Revenue and performance analytics
    - **Compliance Reports**: PCI DSS, GDPR, and SOX reporting
    - **Data Export**: Comprehensive data export capabilities
    
    ### 🔔 Webhook Integration
    - **Real-time Notifications**: Instant transaction updates
    - **Event Processing**: Comprehensive event handling
    - **Retry Logic**: Guaranteed delivery with exponential backoff
    
    ## 🔐 Security & Authentication
    
...
    
...
    
...
    
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /subscriptions/{subscriptionId}:
    get:
      tags:
        - Subscription Management
      summary: Get subscription details
      description: Retrieve detailed information about a specific subscription.
      parameters:
        - name: subscriptionId
          in: path
          required: true
          description: Unique subscription identifier
          schema:
            type: string
      responses:
        '200':
          description: Subscription details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Subscription Management
      summary: Update subscription
      description: Update subscription details such as payment method or billing address.
      parameters:
        - name: subscriptionId
          in: path
          required: true
          description: Unique subscription identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
      responses:
        '200':
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
...
  schemas:
...
    UpdateSubscriptionRequest:
      type: object
      properties:
        paymentMethodId:
          type: string
          format: uuid
        metadata:
          type: object
          additionalProperties: true
...
    SubscriptionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        subscriptionId:
          type: string
          example: sub_1234567890
        customerId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SubscriptionStatus'
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
        nextBillingDate:
          type: string
          format: date-time
        trialStart:
          type: string
          format: date-time
        trialEnd:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SubscriptionStatus:
      type: string
      enum:
        - ACTIVE
        - PAUSED
        - CANCELLED
        - EXPIRED
        - PAST_DUE
        - PENDING
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error category
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
          example: Invalid payment method
        details:
          type: string
          description: Detailed error information
        correlationId:
          type: string
          description: Request correlation ID for tracing
          example: corr-123e4567-e89b-12d3-a456-426614174000
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint path
          example: /api/v1/payments/purchase
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Transaction not found
            correlationId: corr-123e4567-e89b-12d3-a456-426614174000
            timestamp: "2025-09-10T14:30:00Z"
...